[REFACTOR] BaseRandomDto추가하여 재사용성 강화
