+++ src/main/java/org/atdev/artrip/domain/home/web/dto/ScheduleRandomRequestDto.java
@@ -12,29 +12,9 @@ import java.time.LocalDate;
 @NoArgsConstructor
 @AllArgsConstructor
 @Builder
-public class ScheduleRandomRequestDto {
-
-    @NotNull
-    private Boolean isDomestic;
-
-    private String region;
-    private String country;
+public class ScheduleRandomRequestDto extends BaseRandomRequestDto {
 
     @NotNull
     private LocalDate date;
 
-    @Schema(hidden = true)
-    @AssertTrue(message = "국내 전시는 region 필수(전체 가능), 국외 전시는 country 필수(전체 가능)이며 둘을 동시에 보낼 수 없습니다.")
-    public boolean isDomesticRegionCountryValid() {
-        if (isDomestic == null) return true;
-
-        boolean hasRegion = region != null && !region.isBlank();
-        boolean hasCountry = country != null && !country.isBlank();
-
-        if (isDomestic) {
-            return hasRegion && !hasCountry;
-        } else {
-            return hasCountry && !hasRegion;
-        }
-    }
 }
